BlowFish now works ok under 64bit Linux
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@3826 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/core/BlowFish.cpp
@@ -18,7 +18,7 @@
 #define bf_F(x) (((S(x, 0) + S(x, 1)) ^ S(x, 2)) + S(x, 3))
 #define ROUND(a, b, n) (a.word ^= bf_F(b) ^ bf_P[n])
 
-const unsigned long BlowFish::tempbf_P[BlowFish::bf_N + 2] =
+const uint32 BlowFish::tempbf_P[BlowFish::bf_N + 2] =
 {
   0x243f6a88, 0x85a308d3, 0x13198a2e, 0x03707344,
   0xa4093822, 0x299f31d0, 0x082efa98, 0xec4e6c89,
@@ -27,7 +27,7 @@ const unsigned long BlowFish::tempbf_P[BlowFish::bf_N + 2] =
   0x9216d5d9, 0x8979fb1b,
 };
 
-const unsigned long BlowFish::tempbf_S[4][256] =
+const uint32 BlowFish::tempbf_S[4][256] =
 {
   {0xd1310ba6, 0x98dfb5ac, 0x2ffd72db, 0xd01adfb7,
    0xb8e1afed, 0x6a267e96, 0xba7c9045, 0xf12c7f99,
@@ -287,8 +287,8 @@ const unsigned long BlowFish::tempbf_S[4][256] =
    0xb74e6132, 0xce77e25b, 0x578fdfe3, 0x3ac372e6},
 };
 
-void BlowFish::Blowfish_encipher(unsigned long *xl,
-                                 unsigned long *xr)
+void BlowFish::Blowfish_encipher(uint32 *xl,
+                                 uint32 *xr)
 {
   union aword Xl;
   union aword Xr;
@@ -311,8 +311,8 @@ void BlowFish::Blowfish_encipher(unsigned long *xl,
   *xl = Xr.word;
 }
 
-void BlowFish::Blowfish_decipher(unsigned long *xl,
-                                 unsigned long *xr)
+void BlowFish::Blowfish_decipher(uint32 *xl,
+                                 uint32 *xr)
 {
   union aword Xl;
   union aword Xr;
@@ -340,9 +340,9 @@ void BlowFish::InitializeBlowfish(const unsigned char key[],
 {
   short          i;
   short          j;
-  unsigned long  data;
-  unsigned long  datal;
-  unsigned long  datar;
+  uint32  data;
+  uint32  datal;
+  uint32  datar;
   union aword temp;
 
   j = 0;
@@ -416,8 +416,8 @@ void BlowFish::Encrypt(const unsigned char *in, unsigned char *out)
   for (int x = 0; x < 8; x++)
     out[x] = in[x];
 
-  Blowfish_encipher(reinterpret_cast<unsigned long *>(out),
-    reinterpret_cast<unsigned long *>(out + sizeof(unsigned long)));
+  Blowfish_encipher(reinterpret_cast<uint32 *>(out),
+    reinterpret_cast<uint32 *>(out + sizeof(uint32)));
 }
 
 void BlowFish::Decrypt(const unsigned char *in, unsigned char *out)
@@ -425,8 +425,8 @@ void BlowFish::Decrypt(const unsigned char *in, unsigned char *out)
   for (int x = 0; x < 8; x++)
     out[x] = in[x];
 
-  Blowfish_decipher(reinterpret_cast<unsigned long *>(out),
-    reinterpret_cast<unsigned long *>(out + sizeof(unsigned long)));
+  Blowfish_decipher(reinterpret_cast<uint32 *>(out),
+    reinterpret_cast<uint32 *>(out + sizeof(uint32)));
 }
 
 /*
